Crack ROT
This solution makes use of the function rotateAlphabet from
rotn. We removed the prompt asking for input from the encrypt
and decrypt functions. Instead there is now an argument to the
function that has these values. encrypt and decrypt take an optional
argument for the shifted alphabet. If not provided it defaults to the
standard Caesar Cipher shift.

A new option and function called crack was added to the file.
This function tries all possible shifts of the alphabet and uses
these shifted alphabets to decrypt the provided ciphertext. The
results are printed to the terminal.