Tabula Recta - Starting Position Shift
Instead of generating the classic table starting with 'A'. We allow
the user to provide a value from [0, 25] which determines the starting
letter of the table. Depending on how autokey_table was done, converting
that to this could be very easy or a bit difficult. In this solution we
simply move the make position n, in the generate table function a parameter
which can defaults to zero. We added a simple loop to the main function
which checks to see if the user provided a value between [0, 25]. If the
value is between these bounds the generateTable function will generate a
table with its starting row begining with the character 'A' + n where n
is the value provided by the user.